requirements.txt
# requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.0
sqlalchemy
aiosqlite
sqlmodel
httpx
beautifulsoup4
nats-py
asyncio-nats-client
requests
colorama